Move package.json handling to own file
